+++ src/main/java/net/momostudios/coldsweat/common/temperature/modifier/SoulLampTempModifier.java
@@ -15,7 +15,7 @@ public class SoulLampTempModifier extends TempModifier
 {
 
     @Override
-    public float calculate(Temperature temp, PlayerEntity player)
+    public double calculate(Temperature temp, PlayerEntity player)
     {
         float max = (float) ColdSweatConfig.getInstance().maxHabitable();
         float min = (float) ColdSweatConfig.getInstance().minHabitable();
