+++ src/main/java/net/momostudios/coldsweat/common/temperature/modifier/WeatherTempModifier.java
@@ -7,7 +7,7 @@ import net.momostudios.coldsweat.common.temperature.Temperature;
 public class WeatherTempModifier extends TempModifier
 {
     @Override
-    public float calculate(Temperature temp, PlayerEntity player)
+    public double calculate(Temperature temp, PlayerEntity player)
     {
         float weatherTemp = 0;
         if (player.world.canBlockSeeSky(player.getPosition()))
